refactor(commands): split filetype_factory into package by command
