+++ python/mirage/commands/builtin/filetype_factory/handlers/wc.py
@@ -1,0 +1,50 @@
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+from mirage.commands.builtin.filetype_factory.extensions import _fmt
+from mirage.io.types import IOResult
+
+
+async def _ft_wc(resolve_glob,
+                 read,
+                 module,
+                 accessor,
+                 paths,
+                 *texts,
+                 args_l=False,
+                 w=False,
+                 c=False,
+                 m=False,
+                 index=None,
+                 **kwargs):
+    if not paths:
+        raise ValueError("wc: missing operand")
+    paths = await resolve_glob(accessor, paths, index)
+    p = paths[0]
+    if w or c or m:
+        return None, IOResult(
+            exit_code=1,
+            stderr=f"wc: -w/-c/-m not supported for {_fmt(module)} files".
+            encode(),
+        )
+    try:
+        raw = await read(accessor, p, index)
+        return str(module.wc(raw)).encode(), IOResult(
+            reads={p.strip_prefix: raw}, cache=[p.strip_prefix])
+    except Exception as e:
+        return None, IOResult(
+            exit_code=1,
+            stderr=f"wc: {p.original}: failed to read as {_fmt(module)}: {e}".
+            encode(),
+        )
